+++ StRoot/StTpcDb/StRTpcDimensions.h
@@ -1,6 +1,6 @@
 /***************************************************************************
  *
- * $Id: StRTpcDimensions.h,v 1.9 2000/11/14 22:00:06 genevb Exp $
+ * $Id: StRTpcDimensions.h,v 1.8 2000/02/15 22:21:47 hardtke Exp $
  *
  * Author:  David Hardtke
  ***************************************************************************
@@ -10,9 +10,6 @@
  ***************************************************************************
  *
  * $Log: StRTpcDimensions.h,v $
- * Revision 1.9  2000/11/14 22:00:06  genevb
- * Switched several functions from float to double
- *
  * Revision 1.8  2000/02/15 22:21:47  hardtke
  * Add effective drift distances
  *
@@ -55,29 +52,29 @@ public:
     int   numberOfSectors()     const;
 
   //TPC field cage parameters:
-    double ifcRadius()           const;
-    double ofcRadius()           const;
-    double tpcTotalLength()      const;
+    float ifcRadius()           const;
+    float ofcRadius()           const;
+    float tpcTotalLength()      const;
 
   //TPC wheel parameters:
-    double wheelInnerRadius()    const;
-    double wheelOuterRadius()    const;
-    double wheelThickness()      const;
+    float wheelInnerRadius()    const;
+    float wheelOuterRadius()    const;
+    float wheelThickness()      const;
 
-    double senseGasOuterRadius() const;
-    double tpeaThickness()       const;
+    float senseGasOuterRadius() const;
+    float tpeaThickness()       const;
 
   //TPC cathode parameters:
-    double cathodeInnerRadius()  const;
-    double cathodeOuterRadius()  const;
-    double cathodeThickness()    const; 
+    float cathodeInnerRadius()  const;
+    float cathodeOuterRadius()  const;
+    float cathodeThickness()    const; 
 
   //TPC distances
-    double innerEffectiveDriftDistance()  const;
-    double outerEffectiveDriftDistance()  const;
-    double gatingGridZ()                  const;
-    double zInnerOffset()                 const;
-    double zOuterOffset()                 const;
+    float innerEffectiveDriftDistance()  const;
+    float outerEffectiveDriftDistance()  const;
+    float gatingGridZ()                  const;
+    float zInnerOffset()                 const;
+    float zOuterOffset()                 const;
 
 
     ClassDef(StRTpcDimensions,0)
@@ -93,55 +90,55 @@ inline int StRTpcDimensions::numberOfSectors() const {
 return (*mTpc)[0].numberOfSectors;
 }
 
-inline double StRTpcDimensions::ifcRadius() const {
+inline float StRTpcDimensions::ifcRadius() const {
 return (*mTpc)[0].tpcInnerRadius;
 }
     
-inline double StRTpcDimensions::ofcRadius() const {
+inline float StRTpcDimensions::ofcRadius() const {
 return (*mTpc)[0].tpcOuterRadius;
 }
     
-inline double StRTpcDimensions::tpcTotalLength() const {
+inline float StRTpcDimensions::tpcTotalLength() const {
 return (*mTpc)[0].tpcTotalLength;
 }
 
-inline double StRTpcDimensions::wheelInnerRadius() const {
+inline float StRTpcDimensions::wheelInnerRadius() const {
 return (*mTpc)[0].wheelInnerRadius;
 }
 
-inline double StRTpcDimensions::wheelOuterRadius() const {
+inline float StRTpcDimensions::wheelOuterRadius() const {
 return (*mTpc)[0].wheelOuterRadius;
 }
 
-inline double StRTpcDimensions::wheelThickness() const {
+inline float StRTpcDimensions::wheelThickness() const {
 return (*mTpc)[0].wheelThickness;
 }
 
-inline double StRTpcDimensions::senseGasOuterRadius() const {
+inline float StRTpcDimensions::senseGasOuterRadius() const {
 return (*mTpc)[0].senseGasOuterRadius;
 }
     
-inline double StRTpcDimensions::tpeaThickness() const {
+inline float StRTpcDimensions::tpeaThickness() const {
 return (*mTpc)[0].tpeaThickness; 
 }
 
-inline double StRTpcDimensions::cathodeInnerRadius() const {
+inline float StRTpcDimensions::cathodeInnerRadius() const {
 return (*mTpc)[0].cathodeInnerRadius;
 }
     
-inline double StRTpcDimensions::cathodeOuterRadius() const {
+inline float StRTpcDimensions::cathodeOuterRadius() const {
 return (*mTpc)[0].cathodeOuterRadius;
 }
     
-inline double StRTpcDimensions::cathodeThickness() const {
+inline float StRTpcDimensions::cathodeThickness() const {
 return (*mTpc)[0].cathodeThickness;
 } 
 
-inline double StRTpcDimensions::zInnerOffset() const {
+inline float StRTpcDimensions::zInnerOffset() const {
   return (*mEffGeo)[0].z_inner_offset;
 }
 
-inline double StRTpcDimensions::zOuterOffset() const {
+inline float StRTpcDimensions::zOuterOffset() const {
   return (*mEffGeo)[0].z_outer_offset;
 }
