+++ StRoot/StTpcDb/StTpcDimensionsI.h
@@ -1,6 +1,6 @@
 /***************************************************************************
  *
- * $Id: StTpcDimensionsI.h,v 1.4 2000/11/14 22:00:06 genevb Exp $
+ * $Id: StTpcDimensionsI.h,v 1.3 2000/02/15 22:21:47 hardtke Exp $
  *
  * Author:  David Hardtke
  ***************************************************************************
@@ -10,9 +10,6 @@
  ***************************************************************************
  *
  * $Log: StTpcDimensionsI.h,v $
- * Revision 1.4  2000/11/14 22:00:06  genevb
- * Switched several functions from float to double
- *
  * Revision 1.3  2000/02/15 22:21:47  hardtke
  * Add effective drift distances
  *
@@ -33,29 +30,29 @@ public:
   virtual int   numberOfSectors()     const = 0;
 
   //TPC field cage parameters:
-  virtual double ifcRadius()           const = 0;
-  virtual double ofcRadius()           const = 0;
-  virtual double tpcTotalLength()      const = 0;
+  virtual float ifcRadius()           const = 0;
+  virtual float ofcRadius()           const = 0;
+  virtual float tpcTotalLength()      const = 0;
 
   //TPC wheel parameters:
-  virtual double wheelInnerRadius()    const = 0;
-  virtual double wheelOuterRadius()    const = 0;
-  virtual double wheelThickness()      const = 0;
+  virtual float wheelInnerRadius()    const = 0;
+  virtual float wheelOuterRadius()    const = 0;
+  virtual float wheelThickness()      const = 0;
 
-  virtual double senseGasOuterRadius() const = 0;
-  virtual double tpeaThickness()       const = 0;
+  virtual float senseGasOuterRadius() const = 0;
+  virtual float tpeaThickness()       const = 0;
 
   //TPC cathode parameters:
-  virtual double cathodeInnerRadius()  const = 0;
-  virtual double cathodeOuterRadius()  const = 0;
-  virtual double cathodeThickness()    const = 0; 
+  virtual float cathodeInnerRadius()  const = 0;
+  virtual float cathodeOuterRadius()  const = 0;
+  virtual float cathodeThickness()    const = 0; 
 
   //TPC distances
-  virtual double innerEffectiveDriftDistance()  const = 0;
-  virtual double outerEffectiveDriftDistance()  const = 0;
-  virtual double gatingGridZ()                  const = 0;
-  virtual double zInnerOffset()                 const = 0;
-  virtual double zOuterOffset()                 const = 0;
+  virtual float innerEffectiveDriftDistance()  const = 0;
+  virtual float outerEffectiveDriftDistance()  const = 0;
+  virtual float gatingGridZ()                  const = 0;
+  virtual float zInnerOffset()                 const = 0;
+  virtual float zOuterOffset()                 const = 0;
 
 ClassDef(StTpcDimensionsI,0)
